updated the code in add-user route

+++ WalkingSkeletonApi/Controllers/PhotosController.cs
@@ -27,9 +27,8 @@ namespace WalkingSkeletonApi.Controllers
         }
 
 
-        // this route is also open to users that are not logged-in
+        // this route is only open to logged-in users
         // api/Photos/add-photo?userId=d274fa2f-201a-41f9-8a89-f17c4f07544d
-        [AllowAnonymous]
         [HttpPost("add-photo")]
         public async Task<IActionResult> AddPhoto([FromForm] PhotoUploadDto model, string userId)
         {
@@ -103,14 +102,14 @@ namespace WalkingSkeletonApi.Controllers
         public async Task<IActionResult> GetUserMainPhoto(string userId)
         {
             //check if user logged is the one making the changes - only works for system using Auth tokens
-            //ClaimsPrincipal currentUser = this.User;
-            //var currentUserId = currentUser.FindFirst(ClaimTypes.NameIdentifier).Value;
-            //if (!userId.Equals(currentUserId))
-            //{
-            //    ModelState.AddModelError("Denied", $"You are not allowed to upload photo for another user");
-            //    var result2 = Util.BuildResponse<string>(false, "Access denied!", ModelState, "");
-            //    return BadRequest(result2);
-            //}
+            ClaimsPrincipal currentUser = this.User;
+            var currentUserId = currentUser.FindFirst(ClaimTypes.NameIdentifier).Value;
+            if (!userId.Equals(currentUserId))
+            {
+                ModelState.AddModelError("Denied", $"You are not allowed to upload photo for another user");
+                var result2 = Util.BuildResponse<string>(false, "Access denied!", ModelState, "");
+                return BadRequest(result2);
+            }
 
             var photo = await _photoService.GetUserMainPhotoAsync(userId);
             if (photo == null)
@@ -134,14 +133,14 @@ namespace WalkingSkeletonApi.Controllers
         public async Task<IActionResult> SetMainPhoto(string userId, string publicId)
         {
             //check if user logged is the one making the changes - only works for system using Auth tokens
-            //ClaimsPrincipal currentUser = this.User;
-            //var currentUserId = currentUser.FindFirst(ClaimTypes.NameIdentifier).Value;
-            //if (!userId.Equals(currentUserId))
-            //{
-            //    ModelState.AddModelError("Denied", $"You are not allowed to upload photo for another user");
-            //    var result2 = Util.BuildResponse<string>(false, "Access denied!", ModelState, "");
-            //    return BadRequest(result2);
-            //}
+            ClaimsPrincipal currentUser = this.User;
+            var currentUserId = currentUser.FindFirst(ClaimTypes.NameIdentifier).Value;
+            if (!userId.Equals(currentUserId))
+            {
+                ModelState.AddModelError("Denied", $"You are not allowed to upload photo for another user");
+                var result2 = Util.BuildResponse<string>(false, "Access denied!", ModelState, "");
+                return BadRequest(result2);
+            }
 
             var res = await _photoService.SetMainPhotoAsync(userId, publicId);
             if (!res.Item1)
@@ -161,14 +160,14 @@ namespace WalkingSkeletonApi.Controllers
         public async Task<IActionResult> UnsetMainPhoto(string userId)
         {
             //check if user logged is the one making the changes - only works for system using Auth tokens
-            //ClaimsPrincipal currentUser = this.User;
-            //var currentUserId = currentUser.FindFirst(ClaimTypes.NameIdentifier).Value;
-            //if (!userId.Equals(currentUserId))
-            //{
-            //    ModelState.AddModelError("Denied", $"You are not allowed to upload photo for another user");
-            //    var result2 = Util.BuildResponse<string>(false, "Access denied!", ModelState, "");
-            //    return BadRequest(result2);
-            //}
+            ClaimsPrincipal currentUser = this.User;
+            var currentUserId = currentUser.FindFirst(ClaimTypes.NameIdentifier).Value;
+            if (!userId.Equals(currentUserId))
+            {
+                ModelState.AddModelError("Denied", $"You are not allowed to upload photo for another user");
+                var result2 = Util.BuildResponse<string>(false, "Access denied!", ModelState, "");
+                return BadRequest(result2);
+            }
 
             var res = await _photoService.UnSetMainPhotoAsync(userId);
             if (!res)
@@ -182,20 +181,20 @@ namespace WalkingSkeletonApi.Controllers
         }
 
         // this route is only open to logged-in users
-        // api/Photos/delete-photo?publicId=kbumxhsvbyqyebrtagmk
+        // api/Photos/delete-photo/publicId=kbumxhsvbyqyebrtagmk?userId=d274fa2f-201a-41f9-8a89-f17c4f07544d
         [HttpDelete("delete-photo/{publicId}")]
         [AllowAnonymous]
         public async Task<IActionResult> DeletePhoto(string userId, string publicId)
         {
             //check if user logged is the one making the changes - only works for system using Auth tokens
-            //ClaimsPrincipal currentUser = this.User;
-            //var currentUserId = currentUser.FindFirst(ClaimTypes.NameIdentifier).Value;
-            //if (!userId.Equals(currentUserId))
-            //{
-            //    ModelState.AddModelError("Denied", $"You are not allowed to upload photo for another user");
-            //    var result2 = Util.BuildResponse<string>(false, "Access denied!", ModelState, "");
-            //    return BadRequest(result2);
-            //}
+            ClaimsPrincipal currentUser = this.User;
+            var currentUserId = currentUser.FindFirst(ClaimTypes.NameIdentifier).Value;
+            if (!userId.Equals(currentUserId))
+            {
+                ModelState.AddModelError("Denied", $"You are not allowed to upload photo for another user");
+                var result2 = Util.BuildResponse<string>(false, "Access denied!", ModelState, "");
+                return BadRequest(result2);
+            }
 
             var res = await _photoService.DeletePhotoAsync(publicId);
             if (!res)
